Fix proper return value typing around async methods
Make sure async methods have their return values reboxed to the proper structural or nominal types. We change the struct to always store the boxed value, and add a generated $apply method to structs and nominal types that changes raw data (boxed to "Object") into the the well-typed boxed data.