Add the --vcs-dev-dir flag to allow for loading VCS packages locally
When the --vcs-dev-dir flag is specified, and a package is imported without specification (tag, commit or branch), the VCS package loader will now first check for the package in one of the specified development directories. If found, then the local development copy will be used in place of VCS loading. This feature allows for development of multiple dependent projects simultaneously on the same system.