Add the concept of a simple directional iterator
This change causes the most common queries from a single name, over a single predicate, in a single direction, to bypass the normal Cayley iterator system in favor of a more direct approach that only creates a single iterator type. This results in a ~25% speedup, and should result in far less GC as well, as we only create a single iterator object vs multiple levels.