Add graph and generator support for the `struct` type reference
Following this change, a type reference of `struct` can be used to refer to any type that is structural or serializable. This is useful for parameters on async functions, generic fields in structs, and generic constraints that need to be structural for the former two cases. This change also simplifies the EnsureStructural check, making it safer in the process.