Reduce the noise of errors by skipping checks on invalid scoped names
Before this change, if a variable or other named scope was invalid, it would be treated as a value of type `any`, which resulted in resolution errors. After this change, the error of the scope propagates to all sites *using* that scope as well, to ensure we skip that checking path.