(Major Change) Change generator to only promising when necessary
Promises in some browsers can be *incredibly* slow, as they can break the optimizer, are sometimes implemented in non-native code, and add overhead. This change adds tracking for which functions must be promising, which *might* be promising and which definitely are not. This tracking (via the PML scope labels) is then used by the ES5 generator to generate functions and their calls to return promises only if necessary, reducing the above issues significantly.

Note that the promising checking is (of course) not perfect, so the promise labeling mechanism is conservative and will label as `MAYBE` if it isn't 100% sure. Function calls to `MAYBE` functions are treated as promising, with special handling to auto-wrap them in a promise if they don't actually promise. This is slower, but safe.

Future possible improvements include better tracking of "dynamic" promising state (->, anonymous closure calls, etc), faster handling of the `maybe` calls, and perhaps custom Promise implementation for efficiency.