Statically verify agent construction location
Currently, agents can be constructed pretty much anywhere, which breaks the requirement that their `principal` back-reference is set immediately (which is done in the constructor of the composing type). As a result, this principal reference can currently be `null` which *breaks the type system guarantee*. Rather than making the `principal` reference nullable (which would make everything perfectly safe, but result in a degradation of code quality), this change restricts the cases in which an agent can be constructed to code in which we know the back-reference will be set immediately.

If we find in the future this code is insufficient or too limiting, we can revisit this decision, but at first design it appears to address all the common use cases of the new agent type kind, without forcing use to loosen the type on `principal`.