Add basic error handling to the parser code
When the parser now tries to consume a token that does not match that expected, it will call the error production stack to consume tokens until told to cease. This allows the parser to (somewhat) recover from the bad input and try to continue parsing the remainder of the seru file in a somewhat-valid way. The intention is to use this for parsing code under active editing when called via an IDE.