Change diff engine to use type reference adoption for type comparison
Instead of comparing using human-readable strings (which was always a hack), we now "adopt" a type reference from one type graph into the other, replacing all concrete referred types with the equivalent types found in the other graph with the *same name* under the *same package*. If such a type is not found, the types are considered different and the adoption fails.

This will also open up the ability to support subtype-checking in a followup PR, if necessary.