Fix type completion in Grok
This change adds a new function to the parser called `IsTypePrefix`, which makes use of the existing parsing code to determine if the next token in a snippet of code is within a type reference. If so, we return true, and Grok can use this as a signal to display type completion.

Since this reuses the existing parsing infrastructure, the code is a bit wonky, but it has the nice side effect of making any changes to the syntax automatically reflected in Grok's understanding